Move `InMemoryFeedStore` to a shared scope in a separate file
