+++ EssentialApp/EssentialAppTests/FeedAcceptanceTests.swift
@@ -37,7 +37,7 @@ class FeedAcceptanceTests: XCTestCase {
     func test_onLaunch_displaysEmptyFeedWhenCustomerHasNoConnectivityAndNoCache() {
         let feed = launch(httpClient: .offline, store: .empty)
         feed.simulateAppearance()
-
+        
         XCTAssertEqual(feed.numberOfRenderedFeedImageViews(), 0)
     }
     
@@ -74,50 +74,6 @@ class FeedAcceptanceTests: XCTestCase {
     private func enterBackground(with store: InMemoryFeedStore) {
         let sut = SceneDelegate(httpClient: HTTPClientStub.offline, store: store)
         sut.sceneWillResignActive(UIApplication.shared.connectedScenes.first!)
-    }
-    
-    private class InMemoryFeedStore: FeedStore, FeedImageDataStore {
-        private(set) var feedCache: CachedFeed?
-        private var feedImageDataCache: [URL: Data] = [:]
-        
-        private init(feedCache: CachedFeed? = nil) {
-            self.feedCache = feedCache
-        }
-        
-        func deleteCachedFeed(completion: @escaping FeedStore.DeletionCompletion) {
-            feedCache = nil
-            completion(.success(()))
-        }
-        
-        func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping FeedStore.InsertionCompletion) {
-            feedCache = CachedFeed(feed: feed, timestamp: timestamp)
-            completion(.success(()))
-        }
-        
-        func retrieve(completion: @escaping FeedStore.RetrievalCompletion) {
-            completion(.success(feedCache))
-        }
-        
-        func insert(_ data: Data, for url: URL, completion: @escaping (FeedImageDataStore.InsertionResult) -> Void) {
-            feedImageDataCache[url] = data
-            completion(.success(()))
-        }
-        
-        func retrieve(dataForURL url: URL, completion: @escaping (FeedImageDataStore.RetrievalResult) -> Void) {
-            completion(.success(feedImageDataCache[url]))
-        }
-        
-        static var empty: InMemoryFeedStore {
-            InMemoryFeedStore()
-        }
-        
-        static var withExpiredFeedCache: InMemoryFeedStore {
-            InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date.distantPast))
-        }
-        
-        static var withNonExpiredFeedCache: InMemoryFeedStore {
-            InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date()))
-        }
     }
     
     private func response(for url: URL) -> (Data, HTTPURLResponse) {
